telegram_music_bot/requirements.txt
python-telegram-bot[job-queue]==21.10
aiohttp==3.11.13
aiosqlite==0.20.0
python-dotenv==1.0.1
yt-dlp==2025.1.26
pydantic==2.10.6
pytest==8.3.5
pytest-asyncio==0.25.3
